MxShop/requirements.txt
mysqlclient==1.3.12
django==1.11.6
markdown==2.6.11
django-filter==2.0.0
djangorestframework=3.8.2
coreapi=2.3.3
django-guardian=1.4.9
django-crispy-form=1.7.3